Prefix things and error handling

+++ commands/Misc/setprefix.js
@@ -9,14 +9,28 @@ module.exports = {
     usage: "",
     permissions: "ADMINISTRATOR",
     exec: async (client, message, args) => {
+        let newPrefix = args[0]
+
+        if (newPrefix.toLowerCase() === '-default') newPrefix = 't!'
+
+        if (!newPrefix || newPrefix.length > 5) {
+            return message.channel.send('Prefix must be under 5 characters!')
+        }
         if (!message.member.permissions.has("ADMINISTRATOR")) return message.reply("You dont have the required permissions to make this change!").then(msg => {
             setTimeout(() => msg.delete(), 5000)})
-        if (!args[0]) return message.reply("Couldn't register valid characters.")
+
+        if (!newPrefix) return message.reply("Couldn't register valid characters.")
+
+        if (newPrefix === 't!') {
+            await message.guild.me.setNickname(``)
+        } else {
+            await message.guild.me.setNickname(`[${newPrefix}] Tentro`)
+        }
 
 
 
-        setPrefix(message.guild.id, (args[0]))
-        message.reply("Successfully changed prefix!")
+        setPrefix(message.guild.id, newPrefix)
+        message.reply(`Successfully set the prefix to \`${newPrefix}\``)
 
     }
